+++ libs/components/forms/src/lib/modules/checkbox/checkbox-label-text-label.component.ts
@@ -1,0 +1,24 @@
+import { Component, Input } from '@angular/core';
+
+/**
+@internal
+ */
+@Component({
+  selector: 'sky-checkbox-label-text-label',
+  templateUrl: './checkbox-label-text-label.component.html',
+})
+export class SkyCheckboxLabelTextLabelComponent {
+  @Input()
+  public labelHidden = false;
+
+  @Input()
+  public set labelText(value: string) {
+    this.#_labelText = value.trim();
+  }
+
+  public get labelText(): string {
+    return this.#_labelText;
+  }
+
+  #_labelText = '';
+}